% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebola_data.R
\docType{data}
\name{ebola_models}
\alias{ebola_models}
\title{Transmission models extracted in the systematic review of articles related
to Ebola virus disease.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 294 rows and 12 columns.
}
\source{
Nash, Rebecca K., Bhatia Sangeeta, Morgenstern Christian, Doohan
Patrick, Jorgensen David, McCain Kelly, McCabe Ruth et al. "Ebola virus disease
mathematical models and epidemiological parameters:a systematic review."
The Lancet Infectious Diseases 24, no. 12 (2024): e762-e773.
https://doi.org/10.1016/S1473-3099(24)00374-8
}
\usage{
ebola_models
}
\description{
This data set provides all extracted data for mathematical models applied to
Ebola virus disease (EVD).
}
\keyword{datasets}
