% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassa_data.R
\docType{data}
\name{lassa_articles}
\alias{lassa_articles}
\title{Data on the articles identified and included in the systematic review of
articles related to Lassa fever.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 193 rows and 24 columns.
}
\source{
Doohan P, Jorgensen D, Naidoo TM, McCain K, Hicks JT, McCabe R, Bhatia S, Charniga K, Cuomo-Dannenburg G, Hamlet A, Nash RK, Nikitin D, Rawson T, Sheppard RJ, Unwin HJT, van Elsland S, Cori A, Morgenstern C, Imai-Eaton N; Pathogen Epidemiology Review Group. Lassa fever outbreaks, mathematical models, and disease parameters: a systematic review and meta-analysis. Lancet Glob Health. 2024 Dec;12(12):e1962-e1972. doi: 10.1016/S2214-109X(24)00379-6. PMID: 39577970.
}
\usage{
lassa_articles
}
\description{
This data set provides the details of all extracted articles included in
the systematic review for Lassa fever.
}
\keyword{datasets}
