% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_short_names.R
\docType{data}
\name{params_short_names}
\alias{params_short_names}
\title{Programming friendly names for parameters in the params dataset}
\format{
A data frame with 107 rows and 2 columns:
\itemize{
\item parameter_type_short
\item parameter_type_full
}
}
\usage{
params_short_names
}
\description{
The list of parameters in our bespoke database used for data extraction are
not programming-friendly. For instance, one of the parameters is called
"Human delay - Admission to Care/Hospitalisation to Death", which is unwieldy
if you want to filter on this value. The \code{params_short_names} dataset provides
a mapping of the original names to more compact names. For instance, a compact
name for the same parameter is \code{delay_admission_to_death}. These names are
available as a column in the \code{params} dataset for each pathogen when the
datasets are accessed through the companion package
\href{https://github.com/mrc-ide/epireview}{epireview}. It has been provided here
for convenience. Users of \code{epireviewdb} might might prefer to work with
these compact names while operating on the datasets.
}
\keyword{datasets}
