% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sars_data.R
\docType{data}
\name{sars_params}
\alias{sars_params}
\title{Parameters extracted in the systematic review of articles
related to SARS.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 761 rows and 64 columns.
}
\source{
Severe acute respiratory syndrome (SARS) mathematical models and disease parameters: a systematic review and meta-analysis
Christian Morgenstern, Thomas Rawson, Isobel Routledge, Mara Kont, Natsuko Imai-Eaton, Janetta Skarp, Patrick Doohan, Kelly McCain, Rob Johnson, H. Juliette T. Unwin, Tristan Naidoo, Dominic P Dee, Kanchan Parchani, Bethan N Cracknell Daniels, Anna Vicco, Kieran O. Drake, Paula Christen, Richard J Sheppard, Sequoia I Leuba, Joseph T Hicks, Ruth McCabe, Rebecca K Nash, Cosmo N Santoni, Pathogen Epidemiology Review Group, Gina Cuomo-Dannenburg, Sabine van Elsland, Sangeeta Bhatia, Anne Cori
medRxiv 2024.08.13.24311934; doi: https://doi.org/10.1101/2024.08.13.24311934
}
\usage{
sars_params
}
\description{
This data set provides all extracted parameters for Severe Acute Respiratory Syndrome (SARS).
}
\keyword{datasets}
